Change annotation on property `classBaseDirectory` to comply with Gradle 7.
Fixes #143

Annotation `@Input` is no longer allowed on properties of type `File`. Instead, a new method is added that returns the directory path as a String and is annotated with `@Input`. This ensures that up-to-date checks consider the path (but not its content) of the property, as was the behaviour before. The existing method is instead annotated as `@Internal`.