Upgrade Antlr to the latest (4.11.1)
In our project which uses rocker, one of our other dependencies pulls in a later version of Antlr than the 4.5.3 used in Rocker.

It used to pull in 4.8-1 and everything still worked, but since the move to 4.11.1, the Rocker code generation fails with errors such as

```
> java.io.InvalidClassException: org.antlr.v4.runtime.atn.ATN; Could not deserialize ATN with version 3 (expected 4).
  > org.antlr.v4.runtime.atn.ATN; Could not deserialize ATN with version 3 (expected 4).
```

This PR updates the antlr version to 4.11.1, and fixes the one error message that has changed in tests.

I cannot currently get the gradle-plugin module to build, but I'm not sure if this was caused by me 🤔